Feature couple dday

+++ src/main/java/com/ss/heartlinkapi/admin/controller/AdminCoupleController.java
@@ -4,6 +4,8 @@ import com.ss.heartlinkapi.admin.service.AdminCoupleService;
 import com.ss.heartlinkapi.linkmatch.entity.LinkMatchEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -15,29 +17,72 @@ public class AdminCoupleController {
 
     // 매치 질문 조회
     @GetMapping("/questions")
-    public Page<LinkMatchEntity> getAllquestions
-            (@RequestParam(defaultValue = "1") int page, @RequestParam(defaultValue = "10") int size) {
-
-        System.out.println(adminCoupleService.findAllByOrderByIdDesc(page, size).toString());
-        return adminCoupleService.findAllByOrderByIdDesc(page, size);
+    public ResponseEntity<?> getAllquestions
+    (@RequestParam(defaultValue = "1") int page, @RequestParam(defaultValue = "10") int size) {
+        try {
+            Page<LinkMatchEntity> questions = adminCoupleService.findAllByOrderByIdDesc(page, size);
+            return ResponseEntity.ok(questions);
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
     }
 
     // 매치 질문 등록
     @PostMapping("/questions")
-    public void addMatchQuestion(@RequestBody LinkMatchEntity questionText) {
-        adminCoupleService.addMatchQuestion(questionText);
+    public ResponseEntity<?> addMatchQuestion(@RequestBody LinkMatchEntity questionText) {
+        try {
+            LinkMatchEntity result = adminCoupleService.addMatchQuestion(questionText);
+
+            if (questionText == null || questionText.getMatch1() == null || questionText.getMatch2() == null
+                    || questionText.getDisplayDate() == null || questionText.getLinkMatchId() == null) {
+                return ResponseEntity.badRequest().build();
+            }
+            return ResponseEntity.ok(result);
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
     }
 
     // 매치 질문 수정
     @PutMapping("/questions/{questionId}/update")
-    public void updateMatchQuestion(@PathVariable Long questionId, @RequestBody LinkMatchEntity questionText) {
-        adminCoupleService.updateMatchQuestion(questionId, questionText);
+    public ResponseEntity<?> updateMatchQuestion(@PathVariable Long questionId, @RequestBody LinkMatchEntity questionText) {
+        try{
+            LinkMatchEntity result = adminCoupleService.updateMatchQuestion(questionId, questionText);
+
+            if(questionText == null || questionText.getMatch1() == null || questionText.getMatch2() == null
+            || questionText.getDisplayDate() == null || questionText.getLinkMatchId() == null) {
+                return ResponseEntity.badRequest().build();
+            }
+            if (result == null) {
+                return ResponseEntity.notFound().build();
+            }
+
+            return ResponseEntity.ok(result);
+
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
     }
 
     // 매치 질문 삭제
     @DeleteMapping("/questions/{questionId}/delete")
-    public void deleteMatchQuestion(@PathVariable Long questionId) {
-        adminCoupleService.deleteMatchQuestion(questionId);
+    public ResponseEntity<?> deleteMatchQuestion(@PathVariable Long questionId) {
+        try{
+            if(questionId == null) {
+                return ResponseEntity.badRequest().build();
+            }
+
+            LinkMatchEntity question = adminCoupleService.findById(questionId);
+
+            if (question == null) {
+                return ResponseEntity.notFound().build();
+            } else {
+                adminCoupleService.deleteMatchQuestion(questionId);
+                return ResponseEntity.ok().build();
+            }
+        } catch (Exception e) {
+            return ResponseEntity.internalServerError().build();
+        }
     }
 
 }
